+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added support for categorical data
 - Added support for arrays of hashes and Rover data frames
+- Added `output` option to `predict` method
 
 ## 0.1.2 (2020-08-11)